Fixed incorrect handling of responseFlags attribute of the OP_REPLY struct
In pymongo.helpers._unpack_response, there was a problem in how the
responseFlags were being parsed.  Rather than checking to see if each
bit was set with a bitwise AND, the == operator was being used.  This
meant that if more than one flag was ever set, then the == would return
False, and the logic would proceed incorrectly.

This has been changed to `response_flag & 2` to check for bit #1.
The value `2` is derived from boolean 0b0010, meaning bit #1.

The actual problem appears to have been introduced in Mongod 1.6 when
bit #3 is always set:

    AwaitCapable:  Set when the server supports the AwaitData Query
    option. If it doesn't, a client should sleep a little between
    getMore's of a Tailable cursor. Mongod version 1.6 supports AwaitData
    and thus always sets AwaitCapable.